YoutuvbeChatbot/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain-core==0.1.47
langchain-google-genai==1.0.6
langchain-community==0.1.16
youtube-transcript-api==1.2.2
python-dotenv==1.0.1
faiss-cpu==1.7.4
